+++ packages/vsdx/lib/src/parser/stylesheet_parser.dart
@@ -124,6 +124,7 @@ class StyleSheetParser {
               'DblUnderline',
               'DoubleStrikethrough',
               'Overline',
+              'Highlight',
               'ColorTrans',
               'Letterspace',
               'Pos',